后端更新:
- ProductType SKU 自动生成 (格式: PT-{factoryId}-{序号})
- 修复 API 字段映射和类型处理
- 添加 Claude 开发规范文档

前端更新:
- 简化产品类型表单 (移除手动输入代码，添加下拉菜单)
- 修复供应商/客户管理界面
- 修复设备监控和告警页面空指针问题
- 添加 Claude skills 自动化测试脚本

+++ backend-java/src/main/java/com/cretas/aims/dto/DepartmentDTO.java
@@ -35,7 +35,7 @@ public class DepartmentDTO {
     @Size(max = 500, message = "描述不能超过500个字符")
     private String description;
 
-    private Integer managerUserId;
+    private Long managerUserId;
 
     private String managerName; // 负责人姓名（关联查询）
 
@@ -54,7 +54,7 @@ public class DepartmentDTO {
     @Size(max = 50, message = "图标名称不能超过50个字符")
     private String icon;
 
-    private Integer createdBy;
+    private Long createdBy;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime createdAt;
